refactor Effect and TimeEffect to core.
