+++ opentimelineio/core/__init__.py
@@ -68,3 +68,7 @@ from .unknown_schema import (
 from .coordinate_space_reference import (
     CoordinateSpaceReference
 )
+from .effect import (
+    Effect,
+    TimeEffect
+)
